requirements.txt
Flask
gunicorn
pymilvus
nltk
requests
tqdm
webvtt-py
numpy
pypdf
python-docx
